requirements.txt
PyYAML==3.11
Fabric==1.10.1
